# openapi.yaml
openapi: 3.1.0
info:
  title: OpenLnk API
  version: 1.0.0-gate0
  description: >
    Contract-first v1 surface. TS client in packages/api-client is generated from this
    file; CI drift gate fails if the committed spec differs from the FastAPI export.
    All mutations require Idempotency-Key. Commitment writes require version.
servers:
  - url: https://api.openlnk.in/v1
security:
  - bearerAuth: []
paths:
...
  /commitments:
    get:
      summary: Unified ledger (cursor pagination, RLS-scoped)
      parameters:
        - { name: state, in: query, schema: { type: string } }
        - { name: context_id, in: query, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommitmentPage' }
    post:
      summary: Create commitment manually
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommitmentCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Commitment' }
  /commitments/{id}:
    patch:
      summary: State transition / edit (optimistic concurrency, OL-002/006)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommitmentPatch' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Commitment' }
        '409':
          description: Stale version or invalid transition
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 64 }
  schemas:
...
    Commitment:
      type: object
      properties:
        id: { type: string, format: uuid }
        context_id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        counterparty_id: { type: string, format: uuid, nullable: true }
        title: { type: string }
        class: { type: string, enum: [fee, schedule, task, payment, custom] }
        amount_paise: { type: integer, nullable: true }
        due_at: { type: string, format: date-time, nullable: true }
        state: { type: string, enum: [proposed, accepted, in_progress, done, broken, cancelled] }
        at_risk: { type: boolean }
        version: { type: integer }
        provenance_kind: { type: string, enum: [message, voice, camera, manual], nullable: true }
        upi_intent_url: { type: string, nullable: true }
    CommitmentCreate:
      type: object
      required: [context_id, owner_id, title, class]
      properties:
        context_id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        counterparty_id: { type: string, format: uuid }
        title: { type: string }
        class: { type: string }
        amount_paise: { type: integer }
        due_at: { type: string, format: date-time }
    CommitmentPatch:
      type: object
      required: [version]
      properties:
        version: { type: integer }
        state: { type: string }
        title: { type: string }
        due_at: { type: string, format: date-time }
    CommitmentPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Commitment' } }
        next_cursor: { type: string, nullable: true }
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }